+++ lib/pacto.rb
@@ -13,6 +13,8 @@ require "pacto/extensions"
 require "pacto/request"
 require "pacto/response_adapter"
 require "pacto/response"
+require "pacto/stubs/built_in"
+require "pacto/stubs/stub_provider"
 require "pacto/instantiated_contract"
 require "pacto/contract"
 require "pacto/contract_factory"
