+++ lib/pacto/stubs/stub_provider.rb
@@ -1,0 +1,11 @@
+module Pacto
+  module Stubs
+    class StubProvider
+      class << self
+        def instance
+          @instance = BuiltIn.new
+        end
+      end
+    end
+  end
+end
