+++ spec/unit/pacto/instantiated_contract_spec.rb
@@ -70,155 +70,18 @@ module Pacto
     end
 
     describe '#stub!' do
-      let(:request) do
-        double({
-          :host => 'http://localhost',
-          :method => method,
-          :path => '/hello_world',
-          :headers => {'Accept' => 'application/json'},
-          :params => {'foo' => 'bar'}
-        })
-      end
-
-      let(:method) { :get }
-
-      let(:response) do
-        double({
-          :status => 200,
-          :headers => {},
-          :body => body
-        })
-      end
-
-      let(:body) do
-        {'message' => 'foo'}
-      end
-
-      let(:stubbed_request) { double('stubbed request') }
+      let(:request) { double :request }
+      let(:response) { double :response, body: double(:body) }
+      let(:stub_provider) { double :stub_provider }
 
       before do
-        WebMock.should_receive(:stub_request).
-          with(request.method, "#{request.host}#{request.path}").
-          and_return(stubbed_request)
-
-        stubbed_request.stub(:to_return).with({
-          :status => response.status,
-          :headers => response.headers,
-          :body => response.body.to_json
-        })
+        Pacto::Stubs::StubProvider.stub(instance: stub_provider)
       end
 
-      context 'when the response body is an object' do
-        let(:body) do
-          {'message' => 'foo'}
-        end
-
-        it 'should stub the response body with a json representation' do
-          stubbed_request.should_receive(:to_return).with({
-            :status => response.status,
-            :headers => response.headers,
-            :body => response.body.to_json
-          })
-
-          stubbed_request.stub(:with).and_return(stubbed_request)
-
-          described_class.new(request, response).stub!
-        end
-      end
-
-      context 'when the response body is an array' do
-        let(:body) do
-          [1, 2, 3]
-        end
-
-        it 'should stub the response body with a json representation' do
-          stubbed_request.should_receive(:to_return).with({
-            :status => response.status,
-            :headers => response.headers,
-            :body => response.body.to_json
-          })
-
-          stubbed_request.stub(:with).and_return(stubbed_request)
-
-          described_class.new(request, response).stub!
-        end
-      end
-
-      context 'when the response body is not an object or an array' do
-        let(:body) { nil }
-
-        it 'should stub the response body with the original body' do
-          stubbed_request.should_receive(:to_return).with({
-            :status => response.status,
-            :headers => response.headers,
-            :body => response.body
-          })
-
-          stubbed_request.stub(:with).and_return(stubbed_request)
-
-          described_class.new(request, response).stub!
-        end
-      end
-
-      context 'a GET request' do
-        let(:method) { :get }
-
-        it 'should use WebMock to stub the request' do
-          stubbed_request.should_receive(:with).
-            with({:headers => request.headers, :query => request.params}).
-            and_return(stubbed_request)
-          described_class.new(request, response).stub!
-        end
-      end
-
-      context 'a POST request' do
-        let(:method) { :post }
-
-        it 'should use WebMock to stub the request' do
-          stubbed_request.should_receive(:with).
-            with({:headers => request.headers, :body => request.params}).
-            and_return(stubbed_request)
-          described_class.new(request, response).stub!
-        end
-      end
-
-      context 'a request with no headers' do
-        let(:request) do
-          double({
-            :host => 'http://localhost',
-            :method => :get,
-            :path => '/hello_world',
-            :headers => {},
-            :params => {'foo' => 'bar'}
-          })
-        end
-
-        it 'should use WebMock to stub the request' do
-          stubbed_request.should_receive(:with).
-            with({:query => request.params}).
-            and_return(stubbed_request)
-          described_class.new(request, response).stub!
-        end
-      end
-
-      context 'a request with no params' do
-        let(:request) do
-          double({
-            :host => 'http://localhost',
-            :method => :get,
-            :path => '/hello_world',
-            :headers => {},
-            :params => {}
-          })
-        end
-
-        it 'should use WebMock to stub the request' do
-          stubbed_request.should_receive(:with).
-            with({}).
-            and_return(stubbed_request)
-          described_class.new(request, response).stub!
-        end
+      it 'delegates the stubbing to the current stub provider' do
+        stub_provider.should_receive(:stub!).with(request, response, response.body)
+        described_class.new(request, response).stub!
       end
     end
-	end
+  end
 end
